@@ -3,7 +3,7 @@
   "type": "project",
   "description": "ImpressCMS is an open source content management system with a focus on security and speed",
   "minimum-stability": "dev",
-  "license": "GPL-2",
+  "license": "GPL-2.0-only",
   "authors": [
     {
       "name": "marcan"
@@ -19,7 +19,24 @@
     }
   ],
   "require": {
-    "php": ">=5.4.8"
+    "php": ">=5.5.0",
+    "simplepie/simplepie": "1.3.1",
+    "ezyang/htmlpurifier": "4.6.*",
+    "tecnickcom/tcpdf": "6.*",
+    "ext-curl": "*",
+    "ext-gd": "*",
+    "ext-json": "*",
+    "ext-mbstring": "*",
+    "ext-pcre": "*",
+    "ext-session": "*"
+  },
+  "suggest": {
+    "ext-iconv": "Iconv is needed to handle conversions between charsets",
+    "ext-xml": "Faster way then with PHP scripts to handle XML documents",
+    "ext-openssl": "Without it OpenID will not work",
+    "ext-bcmath": "Better precistions in math operations",
+    "ext-PDO": "Use it if you would like to use PDO driver for database",
+    "ext-mysql": "Use it if you would like to use old driver for MySQL (only for PHP 5.x)"
   },
   "autoload": {
     "psr-0" : {"" : "libraries/"},
