@@ -74,7 +74,16 @@ if (preg_match_all('|([^/]+)/([^/]+)/([^/]+)(.*)|', $path, $params, PREG_SET_ORD
         (!empty(\icms::$user) && is_object(\icms::$user)) ? \icms::$user->uid : 0
     );
 
-    $icmsConfig['startpage'] = $icmsConfig['startpage'][$group];
+    // added failover to default startpage for the registered users group -- JULIAN EGELSTAFF Apr 3 2017
+	$groups = (!empty(\icms::$user) && is_object(\icms::$user)) ? \icms::$user->getGroups() : array(ICMS_GROUP_ANONYMOUS);
+	if(($icmsConfig['startpage'][$group] == "" OR $icmsConfig['startpage'][$group] == "--") 
+	AND in_array(ICMS_GROUP_USERS, $groups) 
+	AND $icmsConfig['startpage'][ICMS_GROUP_USERS] != "" 
+	AND $icmsConfig['startpage'][ICMS_GROUP_USERS] != "--") {
+		$icmsConfig['startpage'] = $icmsConfig['startpage'][ICMS_GROUP_USERS];
+	} else {
+		$icmsConfig['startpage'] = $icmsConfig['startpage'][$group];
+	}
 
     if (isset($icmsConfig['startpage']) && $icmsConfig['startpage'] != '' && $icmsConfig['startpage'] != '--') {
         $arr = explode('-', $icmsConfig['startpage']);
