@@ -3,7 +3,7 @@
  You may not change or alter any portion of this comment or credits
  of supporting developers from this source code or any supporting source code
  which is considered copyrighted (c) material of the original comment or credit authors.
- 
+
  This program is distributed in the hope that it will be useful,
  but WITHOUT ANY WARRANTY; without even the implied warranty of
  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
@@ -188,4 +188,3 @@ ob_start();
 $content = ob_get_contents();
 ob_end_clean();
 include 'install_tpl.php';
-?>
