@@ -12,9 +12,6 @@
  * @author		Martijn Hertog (AKA wtravel) <martin@efqconsultancy.com>
  * @version		$Id$
  */
-/**
- *
- */
 require_once 'common.inc.php';
 
 if (!defined( 'XOOPS_INSTALL' ) )	exit();
@@ -67,18 +64,6 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 		} else {
 			$content .= _INSTALL_NO_PLUS_MOD;
 		}
-		//Install protector module by default if found.
-		//TODO: Insert Protector installation - leads to blank page as it is now.
-		if (file_exists(ICMS_ROOT_PATH.'/modules/protector/xoops_version.php')) {
-			$content .= xoops_module_install('protector');
-			/*        	include_once "./class/mainfilemanager.php";
-			 $mm = new mainfile_manager("../mainfile.php");
-			 $mm->setRewrite('PROTECTOR1', 'include  XOOPS_TRUST_PATH.\'/modules/protector/include/precheck.inc.php\')' ;
-			 $mm->setRewrite('PROTECTOR2', 'include  XOOPS_TRUST_PATH.\'/modules/protector/include/postcheck.inc.php\')' ;
-
-			 $result = $mm->doRewrite();
-			 $mm->report();*/
-		}
 
 		$tables = array();
 		$content .= "<div style='height:auto;max-height:400px;overflow:auto;'>".$dbm->report()."</div>";
@@ -87,14 +72,19 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 		exit();
 	}
 } else {
+	$langarr = icms_module_Handler::getAvailable();
+	if (count($langarr) == 1 && in_array('system', $langarr)) {
+		$wizard->redirectToPage( '+1' );
+		exit();
+	}
 
 	$content .= '<div>'. _INSTALL_SELECT_MODS_INTRO .'</div>';
 	$content .= '<div class="dbconn_line">';
 	$content .= '<h3>'. _INSTALL_SELECT_MODULES.'</h3>';
 	$content .= '<div id="modinstall" name="install_mods[]">';
-	$langarr = icms_module_Handler::getAvailable();
+	
 	foreach ($langarr as $lang) {
-		if ($lang == 'system' || $lang == 'protector') {
+		if ($lang == 'system') {
 			continue;
 		}
 		$content .= "<div class=\"langselect\" style=\"text-decoration: none;\"><a href=\"javascript:void(0);\" style=\"text-decoration: none;\"><img src=\"../modules/$lang/images/icon_small.png\" alt=\"$lang\" /><br />$lang <br /><input type=\"checkbox\" checked=\"checked\" name=\"install_mods[]\" value=\"$lang\" /></a></div>";
