@@ -14,65 +14,65 @@ defined("ICMS_ROOT_PATH") or die("ImpressCMS root path not defined");
 
 /**
  * Generates META tags
- * 
+ *
  * @package	ICMS\IPF
  * @todo	Properly set visibility of variables - in version 2.0
  */
 class icms_ipf_Metagen {
-	
+
         /**
          * Instance of icms_core_Textsanitizer
-         * 
+         *
          * @var \icms_core_Textsanitizer
          */
 	public $_myts;
 
-	/** 
+	/**
          * Title
-         * 
-         * @var string 
+         *
+         * @var string
          */
 	public $_title = '';
 
-	/** 
+	/**
          * Original title
-         * 
-         * @var string 
+         *
+         * @var string
          */
 	public $_original_title = '';
 
-	/** 
+	/**
          * Keywords
-         * 
-         * @var string 
+         *
+         * @var string
          */
 	public $_keywords = '';
 
-	/** 
+	/**
          * Meta description
-         * 
-         * @var string 
+         *
+         * @var string
          */
 	public $_meta_description = '';
 
-	/** 
+	/**
          * Category path
-         * 
-         * @var string 
+         *
+         * @var string
          */
 	public $_categoryPath = '';
 
-	/** 
+	/**
          * Description
-         * 
-         * @var string 
+         *
+         * @var string
          */
 	public $_description = '';
 
-	/** 
+	/**
          * Min chars
-         * 
-         * @var int 
+         *
+         * @var int
          */
 	public $_minChar = 4;
 
@@ -395,7 +395,7 @@ class icms_ipf_Metagen {
 	public function createMetaTags() {
 		global $xoopsTpl, $xoTheme;
         $this->_keywords = preg_replace("(<!--.*?-->)", '', $this->_keywords); // stops html comments appearing in meta key
-        
+
 		if (is_object($xoTheme)) {
 			$xoTheme->addMeta('meta', 'keywords', $this->_keywords);
 			$xoTheme->addMeta('meta', 'description', $this->_description);
@@ -408,4 +408,3 @@ class icms_ipf_Metagen {
 	}
 
 }
-
