@@ -420,7 +420,7 @@ class icms_module_Handler
 			if (curl_exec($ch)) {
 				icms_core_Message::error($url . $fields_string, 'Notification Sent to');
 			} else {
-				throw new Execption("Unable to contact update server");
+				throw new Exception("Unable to contact update server");
 			}
 
 			//close connection
