@@ -15,290 +15,290 @@ abstract class icms_properties_Handler implements Serializable {
      * Specifies property that a property stores a string
      */
     const DTYPE_STRING = 2; // XOBJ_DTYPE_TXTBOX
-    
+
     /**
      * Specifies property that a property stores a integer
      */
     const DTYPE_INTEGER = 3; // XOBJ_DTYPE_INT
-    
+
     /**
      * Specifies property that a property stores a float number
      */
     const DTYPE_FLOAT = 201; // XOBJ_DTYPE_FLOAT
-    
+
     /**
      * Specifies property that a property stores a boolean
      */
     const DTYPE_BOOLEAN = 105;
-    
+
     /**
      * Specifies property that a property stores a file
      */
-    const DTYPE_FILE = 104; // 
-    
+    const DTYPE_FILE = 104; //
+
     /**
      * Specifies property that a property stores a date or/and time
      */
     const DTYPE_DATETIME = 11; // XOBJ_DTYPE_LTIME
-    
+
     /**
      * Specifies property that a property stores an array
      */
     const DTYPE_ARRAY = 6; // XOBJ_DTYPE_ARRAY
-    
+
     /**
      * Specifies property that a property stores a list
      */
     const DTYPE_LIST = 101; // XOBJ_DTYPE_SIMPLE_ARRAY
-    
+
     /**
      * Specifies property that a property stores a object
      */
     const DTYPE_OBJECT = 12;
-    
+
     /**
      * Specifies property that a property stores a unknown format data
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_OTHER = 7; // XOBJ_DTYPE_OTHER
-    
+
     /**
      * Specifies property that a property stores a file (old format)
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_FILE = 204; //XOBJ_DTYPE_FILE
-    
+
     /**
      * Specifies property that a property stores a long string)
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_TXTBOX = 1; // XOBJ_DTYPE_TXTBOX
-    
+
     /**
      * Specifies property that a property stores a url string
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_URL = 4; // XOBJ_DTYPE_URL
-    
+
     /**
      * Specifies property that a property stores a email address
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_EMAIL = 5; // XOBJ_DTYPE_EMAIL
-    
+
     /**
      * Specifies property that a property stores a source code
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_SOURCE = 8; // XOBJ_DTYPE_SOURCE
-    
+
     /**
      * Specifies property that a property stores a short time
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_STIME = 9; // XOBJ_DTYPE_STIME
-    
+
     /**
      * Specifies property that a property stores a middle time
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_MTIME = 10; // XOBJ_DTYPE_MTIME
-    
+
     /**
      * Specifies property that a property stores a currency
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_CURRENCY = 200; // XOBJ_DTYPE_CURRENCY
-    
+
     /**
      * Specifies property that a property stores a time only data
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_TIME_ONLY = 202; // XOBJ_DTYPE_TIME_ONLY
-    
+
     /**
      * Specifies property that a property stores a urllink
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_URLLINK = 203; // XOBJ_DTYPE_URLLINK
-    
+
     /**
      * Specifies property that a property stores a image
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_IMAGE = 205; // XOBJ_DTYPE_IMAGE
-    
+
     /**
      * Specifies property that a property stores a form section opening
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_FORM_SECTION = 210; // XOBJ_DTYPE_FORM_SECTION
-    
+
     /**
      * Specifies property that a property stores a form section closing
-     * 
+     *
      * @deprecated since version 2.0
      */
     const DTYPE_DEP_FORM_SECTION_CLOSE = 211; // XOBJ_DTYPE_FORM_SECTION_CLOSE
-        
+
     /**
      * Specifies allowed mimetypes for the var (only suported for DTYPE_FILE)
      */
     const VARCFG_ALLOWED_MIMETYPES = 'allowedMimeTypes';
-    
+
     /**
      * Specifies max filesize for the var (only suported for DTYPE_FILE)
      */
     const VARCFG_MAX_FILESIZE = 'maxFileSize';
-    
+
     /**
      * Specifies max width of image for the var (only suported for DTYPE_FILE)
      */
     const VARCFG_MAX_WIDTH = 'maxWidth';
-    
+
     /**
      * Specifies max height for the var (only suported for DTYPE_FILE)
      */
     const VARCFG_MAX_HEIGHT = 'maxHeight';
-    
+
     /**
      * Specifies prefix for the var (only suported for DTYPE_FILE, DTYPE_STRING)
      */
     const VARCFG_PREFIX = 'prefix';
-    
+
     /**
      * Specifies saving path for the var (only suported for DTYPE_FILE)
      */
     const VARCFG_PATH = 'path';
-    
+
     /**
      * Specifies filename generator function for the var (only suported for DTYPE_FILE)
      */
     const VARCFG_FILENAME_FUNCTION = 'filenameGenerator';
-    
+
     /**
      * Specifies list with options that can be used for the var
      */
     const VARCFG_POSSIBLE_OPTIONS = 'possibleOptions';
-    
+
     /**
      * Specifies if this var is locked from changing?
      */
     const VARCFG_LOCKED = 'locked';
-    
+
     /**
      * Specifies if this var is hidden from user
      */
     const VARCFG_HIDE = 'hide';
-    
+
     /**
      * Specifies how var should be rendered
      */
     const VARCFG_RENDER_TYPE = 'renderType';
-    
+
     /**
      * Specifies separator for saving list items for the var (only suported for DTYPE_LIST)
      */
     const VARCFG_SEPARATOR = 'separator';
-    
+
     /**
      * Specifies max text length for the var (only suported for DTYPE_STRING)
      */
     const VARCFG_MAX_LENGTH = 'maxLength';
-    
+
     /**
      * Specifies validation rule (REGEXP)
      */
     const VARCFG_VALIDATE_RULE = 'validateRule';
-    
+
     /**
      * Specifies source formating options (only usable for DTYPE_STRING)
      */
     const VARCFG_SOURCE_FORMATING = 'sourceFormating';
-    
+
     /**
      * Specifies outputing formating
      */
     const VARCFG_FORMAT = 'format';
-    
+
     /**
      * Is autoforating for this var disabled?
      */
     const VARCFG_AF_DISABLED = 'autoFormatingDisabled';
-    
+
     /**
      * Not GPC?
      */
     const VARCFG_NOT_GPC = 'not_gpc';
-    
+
     /**
      * Is var changed?
      */
     const VARCFG_CHANGED = 'changed';
-    
+
     /**
      * Value of var
      */
     const VARCFG_VALUE = 'value';
-    
+
     /**
      * Var data type
      */
     const VARCFG_TYPE = 'data_type';
-    
+
     /**
      * Is var required?
      */
     const VARCFG_REQUIRED = 'required';
-    
+
     /**
      * Data handler
      */
     const VARCFG_DATA_HANDLER = 'data_handler';
-    
+
     /**
      * Deprecached data type
      */
     const VARCFG_DEP_DATA_TYPE = 'depDataType';
-    
+
     /**
      * Form caption for this var
      */
     const VARCFG_FORM_CAPTION = 'form_caption';
-    
+
     /**
      * Form description for this var
      */
     const VARCFG_FORM_DESC = 'form_dsc';
-    
+
     /**
      * Default value of this var
      */
     const VARCFG_DEFAULT_VALUE = 'default_value';
-    
+
     /**
      * Is var not loaded?
      */
     const VARCFG_NOTLOADED = 'not_loaded';
-    
+
     /**
      * Validation rule for emails
      */
     const VALIDATE_RULE_EMAIL = '/^[a-z0-9]+([_\\.-][a-z0-9]+)*@([a-z0-9]+([\.-][a-z0-9]+)*)+\\.[a-z]{2,}$/i';
-    
+
     /**
      * Validation rule for links
      */
@@ -307,20 +307,20 @@ abstract class icms_properties_Handler implements Serializable {
     /**
      * Vars configuration
      *
-     * @var array 
+     * @var array
      */
-    protected $_vars = array();        
-    
+    protected $_vars = array();
+
     /**
      * Changed vars count
      *
-     * @var int 
+     * @var int
      */
     protected $_changed = 0;
 
     /**
      * Assigns values from array to vars
-     * 
+     *
      * @param array $values     Assoc arary with keys and values to assign
      */
     public function assignVars($values) {
@@ -336,7 +336,7 @@ abstract class icms_properties_Handler implements Serializable {
      * @param	string		$varname            Var name
      * @param	bool            $displayOnForm      Display on form
      * @param	string		$default            Default value
-     * 
+     *
      * @deprecated since version 2.0
      */
     public function initCommonVar($varname, $displayOnForm = true, $default = 'notdefined') {
@@ -349,13 +349,13 @@ abstract class icms_properties_Handler implements Serializable {
         }
         include __DIR__ . '/common/' . $varname . '.php';
         $this->hideFieldFromSingleView($varname);
-    }    
+    }
 
     /**
      * Magic function to get property value by accessing it by name
-     * 
+     *
      * @param string $name
-     * 
+     *
      * @return mixed
      */
     public function __get($name) {
@@ -384,7 +384,7 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Magic function to work with properties as class variables (set them)
-     * 
+     *
      * @param string $name      Var name
      * @param mixed $value      New value
      */
@@ -396,8 +396,8 @@ abstract class icms_properties_Handler implements Serializable {
         if (isset($this->_vars[$name][self::VARCFG_POSSIBLE_OPTIONS]) && !in_array($value, $this->_vars[$name][self::VARCFG_POSSIBLE_OPTIONS]))
             return trigger_error('Option not in array for variable ' . get_class($this) . '::$' . $name . ' not found', E_USER_WARNING);
         $clean = $this->cleanVar($name, $this->_vars[$name][self::VARCFG_TYPE], $value);
-                
-        if ($clean == $this->_vars[$name][self::VARCFG_VALUE])
+
+        if ($clean === $this->_vars[$name][self::VARCFG_VALUE])
             return;
         $this->_vars[$name][self::VARCFG_VALUE] = $clean;
         $this->setVarInfo($name, self::VARCFG_CHANGED, true);
@@ -407,9 +407,9 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Checks if var exists
-     * 
+     *
      * @param string $name  Var name
-     * 
+     *
      * @return bool
      */
     public function __isset($name) {
@@ -450,9 +450,9 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Gets var value for displaying
-     * 
+     *
      * @param string $name
-     * 
+     *
      * @return mixed
      */
     public function getVarForDisplay($name) {
@@ -485,24 +485,24 @@ abstract class icms_properties_Handler implements Serializable {
                 return sprintf(isset($this->_vars[$name][self::VARCFG_FORMAT]) ? $this->_vars[$name][self::VARCFG_FORMAT] : '%d', $this->_vars[$name][self::VARCFG_VALUE]);
             case self::DTYPE_BOOLEAN:
                 return $this->_vars[$name][self::VARCFG_VALUE] ? _YES : _NO;
-            case self::DTYPE_FILE: // XOBJ_DTYPE_FILE                    
+            case self::DTYPE_FILE: // XOBJ_DTYPE_FILE
                 return str_replace(array("&amp;", "&nbsp;"), array('&', '&amp;nbsp;'), @htmlspecialchars($this->_vars[$name][self::VARCFG_VALUE], ENT_QUOTES, _CHARSET));
             case self::DTYPE_DATETIME: // XOBJ_DTYPE_LTIME
                 return date(isset($this->_vars[$name][self::VARCFG_FORMAT]) ? $this->_vars[$name][self::VARCFG_FORMAT] : 'r', $this->_vars[$name][self::VARCFG_VALUE]);
             case self::DTYPE_ARRAY: // XOBJ_DTYPE_ARRAY
-                return $this->_vars[$name][self::VARCFG_VALUE];            
+                return $this->_vars[$name][self::VARCFG_VALUE];
             case self::DTYPE_LIST; // XOBJ_DTYPE_SIMPLE_ARRAY
                 return $this->_vars[$name][self::VARCFG_VALUE];//nl2br(implode("\n", $this->_vars[$name][self::VARCFG_VALUE]));
             default:
                 return (string)$this->_vars[$name][self::VARCFG_VALUE];
         }
     }
-    
+
     /**
      * Gets var value for editing
-     * 
+     *
      * @param string $name
-     * 
+     *
      * @return mixed
      */
     public function getVarForEdit($name) {
@@ -513,7 +513,7 @@ abstract class icms_properties_Handler implements Serializable {
             case self::DTYPE_BOOLEAN:
             case self::DTYPE_FILE: // XOBJ_DTYPE_FILE
             case self::DTYPE_DATETIME: // XOBJ_DTYPE_LTIME
-            case self::DTYPE_ARRAY: // XOBJ_DTYPE_ARRAY            
+            case self::DTYPE_ARRAY: // XOBJ_DTYPE_ARRAY
                 return str_replace(array("&amp;", "&nbsp;"), array('&', '&amp;nbsp;'), @htmlspecialchars($this->_vars[$name][self::VARCFG_VALUE], ENT_QUOTES, _CHARSET));
             case self::DTYPE_LIST: // XOBJ_DTYPE_SIMPLE_ARRAY
                 return $this->getVar($name, 'n');
@@ -526,9 +526,9 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Gets var value for form
-     * 
+     *
      * @param string $name
-     * 
+     *
      * @return mixed
      */
     public function getVarForForm($name) {
@@ -551,7 +551,7 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Sets var value
-     * 
+     *
      * @param string $name      Var name
      * @param mixed $value      New value
      * @param array $options    Options to apply when settings values
@@ -583,7 +583,7 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Gets changes vars
-     * 
+     *
      * @return array
      */
     public function getChangedVars() {
@@ -592,13 +592,13 @@ abstract class icms_properties_Handler implements Serializable {
             if (isset($format[self::VARCFG_CHANGED]) && $format[self::VARCFG_CHANGED]) {
                 $changed[] = $key;
             }
-        }        
+        }
         return $changed;
     }
 
     /**
      * If is object variables has been changed?
-     * 
+     *
      * @return bool
      */
     public function isChanged() {
@@ -607,10 +607,10 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Checks if var is set
-     * 
-     * @param int $type         
+     *
+     * @param int $type
      * @param string $key
-     * 
+     *
      * @return boolean
      */
     private function isVarSet($type, $key) {
@@ -634,7 +634,7 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Gets an array with required but not specified vars
-     * 
+     *
      * @return array
      */
     public function getProblematicVars() {
@@ -648,7 +648,7 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Initialize var (property) for the object
-     * 
+     *
      * @param string $key           Var name
      * @param int $dataType         Var data type (use constants DTYPE_* for specifing it!)
      * @param mixed $defaultValue   Default value
@@ -662,7 +662,7 @@ abstract class icms_properties_Handler implements Serializable {
                 $this->_vars[$key][self::VARCFG_POSSIBLE_OPTIONS] = explode('|', $this->_vars[$key][self::VARCFG_POSSIBLE_OPTIONS]);
             }
         } else {
-            $this->_vars[$key] = array();     
+            $this->_vars[$key] = array();
         }
         switch ($dataType) {
             case self::DTYPE_FILE:
@@ -793,7 +793,7 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Gets default values for all vars
-     * 
+     *
      * @return array
      */
     public function getDefaultVars() {
@@ -806,9 +806,9 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Gets mymetype from filename
-     * 
+     *
      * @param string $filename      Filename
-     * 
+     *
      * @return string
      */
     private function getFileMimeType($filename) {
@@ -825,15 +825,15 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Cleans value for var
-     * 
+     *
      * @param string $key       Var name
      * @param string $type      Var type
      * @param string $value     new value
-     * 
+     *
      * @return mixed
      */
     protected function cleanVar($key, $type, $value) {
-        switch ($type) {            
+        switch ($type) {
             case self::DTYPE_OBJECT:
                 if ($value === null || is_object($value)) {
                     return $value;
@@ -859,7 +859,7 @@ abstract class icms_properties_Handler implements Serializable {
                 return ($value == 'yes') || ($value == 'true');
             case self::DTYPE_LIST:
                 if ((array) ($value) === $value)
-                    return $value;                
+                    return $value;
                 if (empty($value))
                     return array();
                 if (is_string($value))
@@ -867,46 +867,39 @@ abstract class icms_properties_Handler implements Serializable {
                 else
                     return array($value);
             case self::DTYPE_FLOAT:
-                if (is_float($value))
-                    return $value;
-                return floatval($value);
+                return (float)$value;
             case self::DTYPE_INTEGER:
-                if (is_int($value))
-                    return $value;
-                return intval($value);
+                return (int)$value;
             case self::DTYPE_ARRAY:
-                if ((array) $value === $value)
+                if (((array)$value) === $value) {
                     return $value;
-                elseif (is_string($value) && !empty($value)) {
+		}
+		if (empty($value)) {
+		    return array();
+		}
+		if (is_string($value)) {
                     if (in_array(substr($value, 0, 1), array('{', '['))) {
                         $ret = json_decode($value, true);
-                        if ($ret !== null)
-                            return $ret;
                     } elseif (substr($value, 0, 2) == 'a:') {
                         $ret = unserialize($value);
-                        if ($ret !== false)
-                            return $ret;
-                    } elseif ($value == '') {
-                        return array();
                     }
-                    return array($value);
-                } elseif (is_null($value) && empty($value))
-                    return array();
-                elseif (!is_object($value))
-                    return array($value);
-                else
-                    return (array) $value;
+                    if (isset($ret) && ($ret !== null)) {
+			return $ret;
+		    }
+                }
+                return (array)$value;
             case self::DTYPE_FILE:
                 if (isset($_FILES[$key])) {
                     $uploader = new icms_file_MediaUploadHandler($this->_vars[$key]['path'], $this->_vars[$key]['allowedMimeTypes'], $this->_vars[$key]['maxFileSize'], $this->_vars[$key]['maxWidth'], $this->_vars[$key]['maxHeight']);
                     if ($uploader->fetchMedia($key)) {
                         if (!empty($this->_vars[$key][self::VARCFG_FILENAME_FUNCTION])) {
                             $filename = call_user_func($this->_vars[$key][self::VARCFG_FILENAME_FUNCTION], 'post', $uploader->getMediaType(), $uploader->getMediaName());
-                            if (!empty($this->_vars[$key]['prefix']))
-                                $filename = $this->_vars[$key]['prefix'] . $filename;
+                            if (!empty($this->_vars[$key][self::VARCFG_PREFIX])) {
+                                $filename = $this->_vars[$key][self::VARCFG_PREFIX] . $filename;
+							}
                             $uploader->setTargetFileName($filename);
-                        } elseif (!empty($this->_vars[$key]['prefix'])) {
-                            $uploader->setPrefix($this->_vars[$key]['prefix']);
+                        } elseif (!empty($this->_vars[$key][self::VARCFG_PREFIX])) {
+                            $uploader->setPrefix($this->_vars[$key][self::VARCFG_PREFIX]);
                         }
                         if ($uploader->upload()) {
                             return array(
@@ -927,11 +920,12 @@ abstract class icms_properties_Handler implements Serializable {
                     if ($uploader->fetchFromURL($value)) {
                         if (!empty($this->_vars[$key][self::VARCFG_FILENAME_FUNCTION])) {
                             $filename = call_user_func($this->_vars[$key][self::VARCFG_FILENAME_FUNCTION], 'post', $uploader->getMediaType(), $uploader->getMediaName());
-                            if (!empty($this->_vars[$key]['prefix']))
-                                $filename = $this->_vars[$key]['prefix'] . $filename;
+                            if (!empty($this->_vars[$key][self::VARCFG_PREFIX])) {
+                                $filename = $this->_vars[$key][self::VARCFG_PREFIX] . $filename;
+							}
                             $uploader->setTargetFileName($filename);
-                        } elseif (!empty($this->_vars[$key]['prefix'])) {
-                            $uploader->setPrefix($this->_vars[$key]['prefix']);
+                        } elseif (!empty($this->_vars[$key][self::VARCFG_PREFIX])) {
+                            $uploader->setPrefix($this->_vars[$key][self::VARCFG_PREFIX]);
                         }
                         if ($uploader->upload()) {
                             return array(
@@ -944,8 +938,9 @@ abstract class icms_properties_Handler implements Serializable {
                     }
                     return null;
                 } elseif (isset($value['filename']) || isset($value['mimetype'])) {
-                    if (!isset($value['filename']) || !isset($value['mimetype']))
+                    if (!isset($value['filename']) || !isset($value['mimetype'])) {
                         return null;
+					}
                     return $value;
                 }
                 return null;
@@ -953,7 +948,7 @@ abstract class icms_properties_Handler implements Serializable {
                 if (is_int($value))
                     return $value;
                 if ($value === null)
-                    return 0;                
+                    return 0;
                 if (is_numeric($value))
                     return intval($value);
                 if (!is_string($value))
@@ -970,7 +965,7 @@ abstract class icms_properties_Handler implements Serializable {
                 }
                 if (isset($this->_vars[$key][self::VARCFG_NOT_GPC]) && !$this->_vars[$key][self::VARCFG_NOT_GPC] && get_magic_quotes_gpc()) {
                     $value = stripslashes($value);
-                }                
+                }
                 if (!empty($this->_vars[$key][self::VARCFG_VALUE]) && isset($this->_vars[$key][self::VARCFG_VALIDATE_RULE]) && !empty($this->_vars[$key][self::VARCFG_VALIDATE_RULE])) {
                     if (!preg_match($this->_vars[$key][self::VARCFG_VALIDATE_RULE], $value)) {
                         trigger_error(sprintf('Bad format for %s var (%s)', $key, $value), E_USER_ERROR);
@@ -1015,7 +1010,7 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Returns properties as key-value array
-     * 
+     *
      * @return array
      */
     public function toArray() {
@@ -1024,23 +1019,23 @@ abstract class icms_properties_Handler implements Serializable {
             if (isset($this->_vars[$name][self::VARCFG_NOTLOADED]) && $this->_vars[$name][self::VARCFG_NOTLOADED]) {
                 continue;
             }
-            
+
             if (is_object($this->_vars[$name][self::VARCFG_VALUE])) {
                 $ret[$name] = serialize($this->_vars[$name][self::VARCFG_VALUE]);
             } else {
                 $ret[$name] = $this->_vars[$name][self::VARCFG_VALUE];
-            }            
+            }
         }
         return $ret;
     }
 
     /**
      * Returns array of vars or only one var (if name specified) with selected info field
-     * 
+     *
      * @param string $key       Var name
      * @param string $info      Var info to get
      * @param mixed $default    Default response
-     * 
+     *
      * @return mixed
      */
     public function getVarInfo($key = null, $info = null, $default = null) {
@@ -1069,14 +1064,14 @@ abstract class icms_properties_Handler implements Serializable {
      */
     public function &getVars() {
         foreach (array_keys($this->_vars) as $key) {
-            $this->_vars[$key][self::VARCFG_DEFAULT_VALUE] = $this->cleanVar($key, $this->_vars[$key][self::VARCFG_TYPE], isset($this->_vars[$key][self::VARCFG_DEFAULT_VALUE]) ? $this->_vars[$key][self::VARCFG_DEFAULT_VALUE] : null);            
+            $this->_vars[$key][self::VARCFG_DEFAULT_VALUE] = $this->cleanVar($key, $this->_vars[$key][self::VARCFG_TYPE], isset($this->_vars[$key][self::VARCFG_DEFAULT_VALUE]) ? $this->_vars[$key][self::VARCFG_DEFAULT_VALUE] : null);
         }
         return $this->_vars;
     }
 
     /**
      * Return array of properties names
-     * 
+     *
      * @return array
      */
     public function getVarNames() {
@@ -1095,38 +1090,38 @@ abstract class icms_properties_Handler implements Serializable {
             $this->setVar($key, $value, $not_gpc);
         }
     }
-    
+
     /**
      * Changes type for var
-     * 
+     *
      * @param string $key   Var name
      * @param int $type     Type
-     * 
+     *
      * @deprecated since version 2.0
      */
     public function setType($key, $type) {
         icms_core_Debug::setDeprecated("setVarInfo", sprintf(_CORE_REMOVE_IN_VERSION, '2.1'));
         $this->setVarInfo($key, self::VARCFG_TYPE, $type);
     }
-    
+
     /**
      * Sets field as required
-     * 
+     *
      * @param string $key   Var name
      * @param bool $is_required     Is required?
-     * 
-     * @deprecated since version 2.0     
+     *
+     * @deprecated since version 2.0
      */
     public function doSetFieldAsRequired($key, $is_required = true) {
          icms_core_Debug::setDeprecated("setVarInfo", sprintf(_CORE_REMOVE_IN_VERSION, '2.1'));
          $this->setVarInfo($key, self::VARCFG_REQUIRED, $is_required);
     }
-    
+
     /**
      * Returns cleaned vars array
-     * 
+     *
      * @return array
-     * 
+     *
      * @deprecated since version 2.0
      */
     public function cleanVars() {
@@ -1136,27 +1131,27 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Sets var info
-     * 
+     *
      * @param string $key       Var name
      * @param string $info      Var option
      * @param mixed $value      Options value
      */
-    public function setVarInfo($key, $info, $value) {        
+    public function setVarInfo($key, $info, $value) {
         if ($key === null) {
             $key = array_keys($this->_vars);
         }
-        
-        if (is_array($key)) {           
+
+        if (is_array($key)) {
             foreach ($key as $k) {
                 $this->setVarInfo($k, $info, $value);
             }
             return;
-        } 
-        
+        }
+
         if (!isset($this->_vars[$key])) {
             return trigger_error('Variable ' . get_class($this) . '::$' . $key . ' not found', E_USER_WARNING);
         }
-        
+
         $this->_vars[$key][$info] = $value;
         switch ($info) {
             case self::VARCFG_TYPE:
@@ -1167,7 +1162,7 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Serialize instance to string
-     * 
+     *
      * @return string
      */
     public function serialize() {
@@ -1177,7 +1172,7 @@ abstract class icms_properties_Handler implements Serializable {
 
     /**
      * Used when using with unserialize function call
-     * 
+     *
      * @param mixed $serialized
      */
     public function unserialize($serialized) {
