@@ -13,7 +13,7 @@ defined("ICMS_ROOT_PATH") or die("ICMS root path not defined");
 
 /**
  * Class to manage a printer friendly page
- * 
+ *
  * @package	ICMS\View
  * @author      The IcmsFactory <www.smartfactory.ca>
  */
@@ -59,8 +59,7 @@ class icms_view_Printerfriendly {
 		$this->_tpl->assign('icms_print_content', $this->_content);
 		$this->_tpl->assign('icms_print_width', $this->_width);
 
-		$current_urls = icms_getCurrentUrls();
-		$current_url = $current_urls['full'];
+		$current_url = icms::$urls['full'];
 
 		$this->_tpl->assign('icms_print_currenturl', $current_url);
 		$this->_tpl->assign('icms_print_url', $this->url);
