@@ -215,7 +215,7 @@ class icms_view_theme_Object {
   m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
   })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
 
-  ga('create', 'UA-'" . $icmsConfigMetaFooter['google_analytics'] . "', 'auto');
+  ga('create', 'UA-" . $icmsConfigMetaFooter['google_analytics'] . "', 'auto');
   ga('send', 'pageview');
 
 </script>";
